4.21*: blogilistan laajennus, step9

+++ osa 4/blogilista/controllers/blogs.js
@@ -46,6 +46,22 @@ blogsRouter.delete('/:id', async (request, response, next) => {
   console.log("oma delete-metodi");
 
   try{
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+     
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })  
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    console.log(`Poistetaan ${blog}:n blogi`);
+
+    const user = await User.findById(decodedToken.id)
+    console.log(`Pyytäjä: ${user}`);
+
+    if ( blog.user.toString() !== decodedToken.id.toString() ){
+      return response.status(403).json({ error: 'user is not owner this data' })  
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch(exception) {
